Renamed Validated to ImmutableValidated
Also updated the BinaryDigit - see changelog
