+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new map functions that take indices into account (#103)
 - Added support for lambda/constant array initialisation (#104)
 - [BREAKING] Changed array orientation meaning in 1D-2D array conversion functions to be more intuitive. ROW and COLUMN now mean the opposite to before. (#106)
+- [BREAKING] Changed Validated type to ImmutableValidated and updated BinaryDigit to hold bytes instead of ints, for space efficiency.
 
 ### Added
 - Support for writing a 1D one-based array to a 2D one-based array (#9)
